ml-service/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
supabase==2.3.4
scikit-learn==1.4.1.post1
scikit-surprise==1.1.3
pandas==2.2.1
numpy==1.26.4
python-dotenv==1.0.1
